// src/screens/CreateAppointmentScreen/interfaces/interfaces.ts
export interface Appointment {
  id: string;
  patientId: string;
  patientName: string;
  doctorId: string;
  doctorName: string;
  date: string;
  time: string;
  specialty: string;
  status: 'pending' | 'confirmed' | 'cancelled';
}

export interface Doctor {
  id: string;
  name: string;
  specialty: string;
  image: string;
}